组扩展属性的查询和增加修改

+++ common/src/main/java/com/dianrong/common/uniauth/common/interfaces/read/IGroupProfileResource.java
@@ -22,6 +22,6 @@ public interface IGroupProfileResource {
   // scenario: get group profile
   @GET
   @Path("{groupId}?profile_id={profileId}")
-  Response<Map<String, Object>> getGroupProfile(@PathParam("groupId") Long groupId,
+  Response<Map<String, Object>> getGroupProfile(@PathParam("groupId") Integer groupId,
       @PathParam("profileId") Long profileId);
 }
